District and taluka apis work

+++ src/district-taluka/district-taluka.controller.ts
@@ -1,8 +1,9 @@
-import { Controller } from '@nestjs/common';
+import { BadRequestException, Controller, NotFoundException } from '@nestjs/common';
 import { MessagePattern, Payload } from '@nestjs/microservices';
 import { DistrictTalukaService } from './district-taluka.service';
 import { CreateDistrictDto } from './dto/create-district.dto';
 import { UpdateDistrictDto } from './dto/update-district.dto';
+import { isValidObjectId } from 'mongoose';
 
 @Controller()
 export class DistrictTalukaController {
@@ -19,17 +20,39 @@ export class DistrictTalukaController {
   }
 
   @MessagePattern('getDistrictById')
-  getDistrictById(@Payload() { districtId }: { districtId: string }) {
-    return this.districtTalukaService.getDistrictById(districtId);
+  async getDistrictById(@Payload() { districtId }: { districtId: string }) {
+    if (!isValidObjectId(districtId)) {
+      throw new BadRequestException(`No district found with the given ID`);
+    }
+
+    const district = await this.districtTalukaService.getDistrictById(districtId);
+    if (!district) {
+      throw new NotFoundException(`No district found with the given ID`);
+    }
+
+    return district;
   }
 
   @MessagePattern('updateDistrict')
-  updateDistrict(@Payload() { districtId, updateDistrictDto }: { districtId: string; updateDistrictDto: UpdateDistrictDto }) {
-    return this.districtTalukaService.updateDistrict(districtId, updateDistrictDto);
+  async updateDistrict(@Payload() { districtId, ...updateDistrictDto }: { districtId: string } & UpdateDistrictDto) {
+    if (!isValidObjectId(districtId)) {
+      throw new BadRequestException(`No district found with the given ID`);
+    }
+
+    const updatedDistrict = await this.districtTalukaService.updateDistrict(districtId, updateDistrictDto);
+
+    if (!updatedDistrict) {
+      throw new NotFoundException('District not found.');
+    }
+
+    return updatedDistrict;
   }
 
   @MessagePattern('deleteDistrict')
-  deleteDistrict(@Payload() { districtId }: { districtId: string }) {
-    return this.districtTalukaService.deleteDistrict(districtId);
+  async deleteDistrict(@Payload() { districtId }: { districtId: string }) {
+    if (!isValidObjectId(districtId)) {
+      throw new BadRequestException('Invalid district ID format.');
+    }
+    return await this.districtTalukaService.deleteDistrict(districtId);
   }
 }
